test: cover forecast normalization edge cases

+++ SimpleWeatherApi/WeatherForecastService.cs
@@ -9,9 +9,9 @@ public sealed class WeatherForecastService : IWeatherForecastService
         ArgumentException.ThrowIfNullOrWhiteSpace(city);
 
         var normalizedCity = city.Trim();
-        var seed = normalizedCity.ToUpperInvariant().Sum(character => character);
-        var temperatureC = (seed % 35) - 5;
-        var condition = Conditions[seed % Conditions.Length];
+        var seed = normalizedCity.ToUpperInvariant().Sum(character => (long)character);
+        var temperatureC = (int)(seed % 35) - 5;
+        var condition = Conditions[(int)(seed % Conditions.Length)];
 
         return new WeatherForecast(normalizedCity, temperatureC, condition);
     }
